+++ java/fury-core/src/main/java/io/fury/codegen/Expression.java
@@ -48,10 +48,10 @@ import com.google.common.base.Preconditions;
 import com.google.common.primitives.Primitives;
 import com.google.common.reflect.TypeToken;
 import io.fury.type.TypeUtils;
-import io.fury.util.Functions;
 import io.fury.util.Platform;
 import io.fury.util.ReflectionUtils;
 import io.fury.util.StringUtils;
+import io.fury.util.function.Functions;
 import java.lang.reflect.Array;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
