+++ java/fury-core/src/main/java/io/fury/codegen/ExpressionUtils.java
@@ -29,8 +29,8 @@ import com.google.common.base.Preconditions;
 import com.google.common.reflect.TypeToken;
 import io.fury.codegen.Expression.Cast;
 import io.fury.codegen.Expression.Null;
-import io.fury.util.Functions;
 import io.fury.util.StringUtils;
+import io.fury.util.function.Functions;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Arrays;
