+++ java/fury-core/src/main/java/io/fury/resolver/ClassInfo.java
@@ -19,8 +19,8 @@ package io.fury.resolver;
 import com.google.common.base.Preconditions;
 import io.fury.Language;
 import io.fury.serializer.Serializer;
-import io.fury.util.Functions;
 import io.fury.util.ReflectionUtils;
+import io.fury.util.function.Functions;
 
 /**
  * This class put together object type related information to reduce array/map loop up when
