+++ java/fury-core/src/main/java/io/fury/resolver/ClassResolver.java
@@ -79,11 +79,11 @@ import io.fury.type.ClassDef;
 import io.fury.type.Descriptor;
 import io.fury.type.GenericType;
 import io.fury.type.TypeUtils;
-import io.fury.util.Functions;
 import io.fury.util.LoggerFactory;
 import io.fury.util.Platform;
 import io.fury.util.ReflectionUtils;
 import io.fury.util.StringUtils;
+import io.fury.util.function.Functions;
 import java.io.Externalizable;
 import java.io.Serializable;
 import java.lang.reflect.Field;
