+++ java/fury-core/src/test/java/io/fury/util/FunctionsTest.java
@@ -16,6 +16,7 @@
 
 package io.fury.util;
 
+import io.fury.util.function.Functions;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.function.Consumer;
